Split off `Value` into its own file, for slightly less code clutter
